Fixed a bug related to source update in Cloud Func
Problem: A dependency on a static name
From the Cloud Function's perspective, its source config is static.
It has no idea that the content of that object in the bucket has been updated.
Therefore, terraform sees no reason to update the function.

Solution: Trach the object's version (generation)
When a new version of an object is uploaded to the bucket, it gets a new, unique generation number.
This number acts as a unique version identifier for the object's content.
Now, the cloud function resource specifically uses the new generation of the object.